simulation: refactors WarpedMap sources
Details:
- Simplifies code structure
- Adapts new imports rules (ie "no relative import")
- WarpedMap now renders already warped data, and no more calls any data
  or method from the store
- SimulationWarpedMap loads everything it needs from the store (current
  path, selected train...), handles warping the data, and gives
  everything to a WarpedMap
